Get ruff fully clean so CI is actually green
- Wrap multi-line implicit string concatenations in explicit parens
  (ISC004) across deviceplans.py, gui.py, and calibrate_session.py.
- Fix the same whitespace-bypass validation bug found earlier in
  report.py's readiness_warnings, duplicated in calibrate_session.py's
  _chain_state_value_missing (a lone-space chain-state value silently
  passed as "filled in").
- Remove a redundant local re-import in test_steps.py that only existed
  to be referenced by a dead "keep import used" line.
- Drop redundant int() around round() (already returns int), an
  executable bit on two scripts nothing runs directly (matching the
  project's dominant tools/ convention), and add noqa with rationale for
  the three findings that are correct as written (a best-effort UI
  autofill swallow, a 3.10-compatible __enter__ return type, and a
  ValueError kept for its except clause instead of TypeError).

+++ src/smallhd_cal/measurement.py
@@ -81,7 +81,7 @@ def _parse_channel(raw: str) -> float:
 
 
 def _float_to_u8(value: float) -> int:
-    return int(round(max(0.0, min(1.0, value)) * 255.0))
+    return round(max(0.0, min(1.0, value)) * 255.0)
 
 
 def _measurement_to_json(measurement: Measurement) -> dict[str, Any]:
